test: 🚚 Move issue #13 variant tests to top-level integration file
Pulls the four variant tests added in 2c4a9b4, 0cad4f8, 45f61fc, and
8eba60c out of the inline mod tests block in queries.rs and into a
dedicated tests/call_site_variants.rs integration test file.

Tests now exercise only the public crate API (extract_all_php_patterns,
parse_php, language_php, EnvMatch) — matching the convention used by
tests/integration_tests.rs.

Behaviour unchanged: all 4 tests still pass; lib test count returns to
its pre-PR baseline of 125 since the tests now run under
`cargo test --test call_site_variants` instead of via `cargo test --lib`.
